refactor(S12P11A404-123): 카테고리 API 수정
- 카테고리명 중복 확인 로직 삭제
- 인가 로직 추가

+++ server/src/main/java/com/kiosk/server/store/controller/dto/FindAllCategoriesResponse.java
@@ -1,4 +1,4 @@
 package com.kiosk.server.store.controller.dto;
 
-public record FindAllCategoriesResponse (int categoryId, String categoryName) {
+public record FindAllCategoriesResponse (long categoryId, String categoryName) {
 }
